+++ python/nav/models/manage.py
@@ -971,7 +971,7 @@ class NetboxType(models.Model):
         if self.sysobjectid.startswith(prefix):
             specific = self.sysobjectid[len(prefix):]
             enterprise = specific.split('.')[0]
-            return long(enterprise)
+            return int(enterprise)
 
 #######################################################################
 ### Device management
